+++ classes/Mvc/View/TwigView.php
@@ -38,12 +38,11 @@ class TwigView extends AbstractView {
   }
 
   /**
-   * Adds a filter to use in the template
-   *
-   * @param $name string Name of the filter to use in the template
-   * @param $function string Name of the function to execute for the value from the template
+   * @param string $name Name of the filter to use in the template
+   * @param mixed $function Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
    */
-  public function add_output_filter($name, $function, $htmlEscape = FALSE) {
+  public function addOutputFilter($name, $function, $htmlEscape = FALSE) {
     $options = [];
     if (!$htmlEscape) {
       $options = ['is_safe' => ['all']];
@@ -55,17 +54,18 @@ class TwigView extends AbstractView {
    * @param $name
    * @return bool
    */
-  public function has_output_filter($name) {
+  public function hasOutputFilter($name) {
     return $this->getRenderingEngine()->getFilter($name) instanceof \Twig_SimpleFilter;
   }
 
   /**
    * Adds a function to use in the template
    *
-   * @param $name string Name of the function to use in the template
-   * @param $function string Name of the function to execute for the value from the template
+   * @param string $name string Name of the function to use in the template
+   * @param mixed $function string Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
    */
-  public function add_output_function($name, $function, $htmlEscape = FALSE) {
+  public function addOutputFunction($name, $function, $htmlEscape = FALSE) {
     $options = [];
     if (!$htmlEscape) {
       $options = ['is_safe' => ['all']];
@@ -74,11 +74,51 @@ class TwigView extends AbstractView {
   }
 
   /**
-   * @param $name
+   * @param string $name
    * @return bool
    */
-  public function has_output_function($name) {
+  public function hasOutputFunction($name) {
     return $this->getRenderingEngine()->getFunction($name) instanceof \Twig_SimpleFunction;
+  }
+
+  /**
+   * @param string $name Name of the filter to use in the template
+   * @param mixed $function Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->addOutputFilter() instead
+   */
+  public function add_output_filter($name, $function, $htmlEscape = FALSE) {
+    $this->addOutputFilter($name, $function, $htmlEscape);
+  }
+
+  /**
+   * @param string $name
+   * @return bool
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->hasOutputFilter() instead
+   */
+  public function has_output_filter($name) {
+    return $this->hasOutputFilter($name);
+  }
+
+  /**
+   * Adds a function to use in the template
+   *
+   * @param string $name string Name of the function to use in the template
+   * @param mixed $function string Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->addOutputFunction() instead
+   */
+  public function add_output_function($name, $function, $htmlEscape = FALSE) {
+    $this->addOutputFunction($name, $function, $htmlEscape);
+  }
+
+  /**
+   * @param string $name
+   * @return bool
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->hasOutputFunction() instead
+   */
+  public function has_output_function($name) {
+    return $this->hasOutputFunction($name);
   }
 
 }